+++ changelog.md
@@ -5,6 +5,7 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Support non-string choices in colored help screens. Closes {issue}`1284`.
 - Fix reconciliation of flags' environment variables.
 - Force requirement on `cloup >= 3.0.7`.